backend/requirement.txt
fastapi
uvicorn[standard]
authlib
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
asyncpg
python-multipart
sqlmodel
itsdangerous
requests
xmltodict
cloudinary
